doc: adjust layout component for the new home page
