+++ docs/components/home-page.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default () => {
+  return <div>Home</div>
+}
