+++ docs/temp-home.zh.md
@@ -1,0 +1,7 @@
+---
+title: Ant Design Mobile
+full: true
+gapless: true,
+---
+
+<code src="./components/home-page.tsx" inline="true"></code>
